ecommerce-chatbot-backend-with-comments/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Faker
PyJWT
Werkzeug
